@@ -1,8 +1,9 @@
 name: kdl
-version: 0.1.0
+version: 1.0.0-dev
 description: Dart implementation of the KDL Document Language
 repository: https://github.com/jellymann/kdl-dart
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 dev_dependencies:
-  test: ^1.15.7
+  test: ^1.19.5
+  path: ^1.8.0
